+++ tomcat/src/main/java/org/apache/coyote/response/StatusCode.java
@@ -4,6 +4,7 @@ public enum StatusCode implements Assemblable {
 
     OK(200),
     FOUND(302),
+    UNAUTHORIZED(401),
     NOT_FOUND(404),
     INTERNAL_SERVER_ERROR(500),
     ;
